28 Fix Can't find data record in course_categories
The default category was set to 1 which is the Miscellaneous category.
However there's no reason that this category must exist and it may be
deleted.  If this category had been deleted then the error "Can't find data
record in database table course_categories" would appear".  So we now
default to $CFG->defaultrequestcategory which is 1 by default but gets
changed to another category if Miscellaneous is deleted.